Create an initial version of email automation tools
This is the initial demo for email automation that detect the difference between two files got downloaded to. Currently, I can simply detect the diff between two files and then send an email using SendGrid to update the receiver about the files. The way to use and what to do in the future is given below